Fit BBC BASIC into an Agon (#14)
zap could not assemble BBC BASIC on real hardware. It reset-looped, and
had done since well before today's work. On the host it assembles
perfectly, which is why 486 assertions, 247 corpus sources and ASan all
passed: the failure is memory exhaustion, and a 512 KB limit does not
exist on a development machine.

Instrumenting every allocation for that source found 407.7 KB live at
peak, against 512 KB shared with MOS and zap's own 53 KB binary. Three
changes bring it down.

The output buffer was a flat 128 KB whatever the program was. BBC BASIC
emits 20,883 bytes and never touched 107 KB of it. pr_reserve already
doubles the buffer on demand, so the fixed start bought nothing but a
quarter of the machine. It starts at 8 KB now and grows: -96 KB.

hash_node held its key inline, so every bucket reserved 64 bytes of name
whether it held a three-character one or nothing at all. Names move into
one block shared by the table and a node holds an offset into it, which
takes the node from 71 bytes to 10 on the eZ80. BBC BASIC's 1,619 symbols
went from about 144 KB of table to about 20 KB.

label_node loses three bytes: an operand's expression is captured into a
128-byte buffer so its length is a byte, and a line number does not need
24 bits. Small, but a large program holds thousands of these -- BBC BASIC
has 2,149 live at once.

  bbcbasicvez.s on an Agon
    before   reset loop, no output
    after    22.86 s, 20,883 bytes, byte-identical to the host's

The key pool was measured earlier at 0.4% slower and set aside as a
memory-for-speed trade. That measurement was taken on big.asm, which has
no labels at all and therefore never allocates a single node: it measured
the cost and none of the benefit. It is not a trade, it is the fix for
programs zap could not assemble.

test/corpus.sh takes real programs as extra arguments now. The reference's
247 tests are all small and single-file -- none nests includes or has more
than a handful of symbols -- and every interesting failure this project
has had came from real sources instead. BBC BASIC and Rokky both assemble
identically to ez80asm.

Also pinned: the longest deferred expression that works, since its length
is now stored in a byte. Refusing anything past 128 characters is a
divergence from the reference, which has no such limit; it is not new and
is left for the compatibility notes.

249 sources compared: 130 identical, 119 rejected by both, 0 disagreed.

+++ src/label_stack.h
@@ -46,12 +46,20 @@ typedef struct _label_node {
      * in nearly every real program. Storing it as text and re-evaluating it
      * once the symbols are known keeps that working in one pass. */
     int text_off_;
-    int text_len_;
+
+    /* An operand's expression is captured into a 128-byte buffer, so its
+     * length is a byte. A whole int here cost two bytes a node on the eZ80,
+     * and a large program holds thousands of these at once: BBC BASIC has
+     * 2,149 live. */
+    uint8_t text_len_;
 
     int bpos_;   /* offset in the output buffer to patch */
     int next_;   /* address of the instruction after this one, for FIX_REL8 */
     int here_;   /* the statement's address, so '$' still means something */
-    int line_;   /* source line, so the error points at the reference */
+
+    /* The source line, for the diagnostic. 24 bits of it is more line numbers
+     * than a file the reader will accept can have. */
+    uint16_t line_;
 
     uint8_t kind_;
     uint16_t scope_;  /* which local scope the names were written in */
